requirements.txt
Django==1.8.2
boto==2.38.0
django-crispy-forms==1.4.0
django-haystack==2.4.0
django-storages-redux==1.3
elasticsearch==1.6.0
gunicorn==19.3.0
tablib==0.10.0
urllib3==1.11
xlrd==0.9.3
